langchain-stock-sentiment/requirements.txt
langchain
langchain-core
langchain-openai
langchain-community
langfuse
fastapi
python-dotenv
requests
pandas
feedparser
uvicorn